refactor(tests): Rename integration_tests.rs to ai_flow_tests.rs
