+++ src/ai/ai_events_tests.rs
@@ -4,12 +4,12 @@
 //! Tests are organized into separate submodules for better organization.
 
 // Re-export test modules
+#[path = "ai_events_tests/ai_flow_tests.rs"]
+mod ai_flow_tests;
 #[path = "ai_events_tests/application_tests.rs"]
 mod application_tests;
 #[path = "ai_events_tests/debounce_tests.rs"]
 mod debounce_tests;
-#[path = "ai_events_tests/integration_tests.rs"]
-mod integration_tests;
 #[path = "ai_events_tests/property_tests.rs"]
 mod property_tests;
 #[path = "ai_events_tests/query_result_tests.rs"]
